Theme A: extract label re-embed orchestration into vtscore.detectors.embedder_sync
Move _maybe_start_label_reembed and its two embedder helpers out of
vtsearch/routes/detectors/registry.py into a new Flask-free library module
vtscore/detectors/embedder_sync.py. The only app-tier dependency, the
thread-spawn helper, is injected by the route so the module stays
import-clean. The load route is now request<->library glue.
