+++ vtsearch/routes/detectors/registry.py
@@ -307,108 +307,6 @@ def register_detector_from_labelset(importer_name: str):  # noqa: C901
 # ---------------------------------------------------------------------------
 
 
-def _active_dataset_embedder_name() -> str:
-    """Return the embedder name recorded on the active dataset's medias, or ``""``."""
-    from vtsearch.state import snapshot_medias
-
-    snap = snapshot_medias()
-    if not snap:
-        return ""
-    first = next(iter(snap.values()), {})
-    return first.get("embedder", "") or ""
-
-
-def _embedder_display_name(embedder_name: str) -> str:
-    """Return a human-friendly name for *embedder_name*, falling back to the id."""
-    if not embedder_name:
-        return ""
-    from vtscore.media import get_embedder
-
-    try:
-        return get_embedder(embedder_name).display_name or embedder_name
-    except KeyError:
-        return embedder_name
-
-
-def _maybe_start_label_reembed(det_ctx, entry: dict) -> str | None:
-    """Fire a re-embed task when the active dataset uses a different embedder.
-
-    Returns the task id when work was started, or ``None`` when the cache is
-    already aligned (same embedder, empty dataset, no labelset cached, etc.)
-    so the caller can return the synchronous fast-path.
-    """
-    new_embedder = _active_dataset_embedder_name()
-    if not new_embedder or not det_ctx.embedder or new_embedder == det_ctx.embedder:
-        return None
-
-    labelset = det_ctx.cached_labelset
-    if labelset is None or not labelset.elements:
-        # Nothing to re-embed; just update the stamp so we don't re-enter
-        # this branch on every subsequent switch.
-        det_ctx.embedder = new_embedder
-        return None
-
-    from vtscore.concurrency.progress import CancelledError, detector_loading_tasks
-    from vtsearch.state import get_active_context
-    from vtscore.state.core import thread_dataset_context, thread_detector_context
-
-    _thread_ds_ctx = get_active_context()
-    media_type = det_ctx.cached_labelset_media_type or entry.get("media_type", "") or ""
-    display = _embedder_display_name(new_embedder)
-    base_msg = f"Re-resolving labels for {display}…" if display else "Re-resolving labels…"
-
-    task_id = f"_detreembed_{det_ctx.detector_id[:8]}"
-    tracker = detector_loading_tasks.create_task(
-        task_id,
-        entry.get("name", det_ctx.detector_id),
-        detector_id=det_ctx.detector_id,
-        media_type=media_type,
-        embedder=new_embedder,
-    )
-    tracker.update("loading", base_msg, 0, 0, step=1, total_steps=1)
-
-    def reembed_task():
-        from vtscore.detectors.labelset_training import train_from_labelset
-
-        try:
-            with thread_dataset_context(_thread_ds_ctx), thread_detector_context(det_ctx):
-
-                def _embed_progress(name: str, done: int, total: int) -> None:
-                    tracker.check_cancelled()
-                    tracker.update("loading", base_msg, done, total, step=1, total_steps=1)
-
-                # ``populate_label_embeddings`` clears the cache when it detects
-                # the embedder change, so this rebuilds against the new embedder
-                # from scratch. The stamp on ``det_ctx.embedder`` is updated
-                # inside ``populate_label_embeddings``.
-                from vtsearch.state import snapshot_medias
-
-                try:
-                    train_from_labelset(
-                        det_ctx,
-                        labelset,
-                        media_type=media_type,
-                        snap=snapshot_medias(),
-                        on_progress=_embed_progress,
-                    )
-                    tracker.update("idle", "", 0, 0, step=None, total_steps=None)
-                except CancelledError:
-                    tracker.update("idle", "", 0, 0, error="Cancelled", step=None, total_steps=None)
-                except Exception as e:
-                    import traceback as _tb
-
-                    _tb.print_exc()
-                    error_msg = str(e) or repr(e) or "Unknown error during label re-embedding"
-                    tracker.update("idle", "", 0, 0, error=error_msg, step=None, total_steps=None)
-        finally:
-            detector_loading_tasks.mark_finished(task_id)
-
-    from vtsearch.threading import spawn
-
-    spawn(reembed_task, name=f"det-reembed-{det_ctx.detector_id[:8]}")
-    return task_id
-
-
 @detectors_registry_bp.route("/api/detectors/registry/load", methods=["POST"])
 @detectors_registry_bp.arguments(DetectorRegistryLoadRequestSchema)
 @detectors_registry_bp.response(200, DetectorRegistryLoadResponseSchema)
@@ -469,9 +367,12 @@ def load_detector_route(body: dict):  # noqa: C901
         # ``populate_label_embeddings``; this branch just makes the work
         # visible via a progress task instead of letting it run lazily
         # inside the next vote or learned-sort request.
+        from vtscore.detectors.embedder_sync import maybe_start_label_reembed
+        from vtsearch.threading import spawn
+
         det_ctx_existing = get_active_detector_context()
         if det_ctx_existing.detector_id == detector_id:
-            task_id = _maybe_start_label_reembed(det_ctx_existing, entry)
+            task_id = maybe_start_label_reembed(det_ctx_existing, entry, spawn=spawn)
             if task_id is not None:
                 return {
                     "ok": True,
